WIP naprawa

+++ oolab/src/main/java/agh/ics/oop/view/util/AnimalElementBox.java
@@ -1,6 +1,6 @@
 package agh.ics.oop.view.util;
 
-import agh.ics.oop.model.mapObjects.Animal;
+import agh.ics.oop.model.MapObjects.Animal;
 import javafx.scene.layout.VBox;
 import javafx.scene.paint.Color;
 import javafx.scene.shape.Circle;
@@ -9,7 +9,7 @@ public class AnimalElementBox extends VBox {
     private final Animal animal;
     private final int initialEnergy;
 
-    public AnimalElementBox(Animal animal, int cellWidth, int cellHeight, int initialEnergy) {
+    public AnimalElementBox(Animal animal, double cellWidth, double cellHeight, int initialEnergy) {
         this.animal = animal;
         this.initialEnergy = initialEnergy;
         this.setPrefSize(cellWidth, cellHeight);
